pyproject.toml
[project]
name = "nuvex"
version = "0.1.0"
description = "Networked Unified Verified Execution — LangGraph-based autonomous agent platform"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.3",
    "langgraph-checkpoint-postgres>=2.0",
    "langchain-core>=0.3",
    "langchain-openai>=0.3",
    "langchain-anthropic>=0.3",
    "langchain-groq>=0.2",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "psycopg[binary,pool]>=3.2",
    "pgvector>=0.3",
    "alembic>=1.14",
    "sqlalchemy>=2.0",
    "python-telegram-bot>=21.0",
    "aiosmtplib>=3.0",
    "aioimaplib>=1.1",
    "apscheduler>=3.10",
    "croniter>=3.0",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "tiktoken>=0.8",
    "python-multipart>=0.0.22",
    "docker>=7.0",
    "langchain-mcp-adapters>=0.2",
    "cryptography>=43.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-httpx>=0.32",
    "anyio>=4.0",
    "ruff>=0.8",
]

[project.scripts]
nuvex = "nuvex.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/brain", "src/shared"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
